requirements.txt
Pillow>=8.1.2
flask
werkzeug
waitress
